images/jobs/draft_fetcher/requirements.txt
dlt==1.9.0
dlt[bigquery]
pymongo
google-cloud-bigquery
google-auth